+++ infrastructure/modules/armonik/polling-agent-configmap.tf
@@ -12,5 +12,7 @@ resource "kubernetes_config_map" "polling_agent_config" {
     InitWorker__WorkerCheckRetries = "10" # TODO: make it a variable
     InitWorker__WorkerCheckDelay   = "00:00:10" # TODO: make it a variable
     Amqp__LinkCredit               = "2"
+    Pollster__GraceDelay           = "00:00:15"
+    Amqp__PartitionId              = (local.default_partition == null || !contains(keys(var.compute_plane), local.default_partition) ? "" : local.default_partition)
   }
 }