+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Added
 
 * Add static partitioning in ArmoniK
 * Add ArmoniK configmaps: compute-plane-configmap and control-plane-configmap
-* Add the environment variable `Amqp__PartitionId` in the of ArmoniK polling agent container
+* Add the environment variable `Amqp__PartitionId` and `Pollster__GraceDelay` in the of ArmoniK polling agent container
 * Add partition metrics exporter service (monitoring)
 * Add a pod to insert partitions data in the database
 * Add partition metrics exporter in Prometheus scraping list